\name{print.cuba}
\alias{print.cuba}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Method print for package R2Cuba }
\description{
  Print an object of class \sQuote{cuba}, i.e an output of the functions
  of this package.
}
\usage{
print.cuba(x,  ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ an object of class \sQuote{cuba}}
   \item{\dots}{ arguments passed to the default function \sQuote{format}}
}
\value{
  No return
}
\seealso{ \code{\link{format}}, \code{\link{R2Cuba-package}} }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ methods }
